// temp_src_upload/src/types.ts
export interface City {
  id: string;
  name: string;
  provinceId: string;
}

export interface Area {
  id: string;
  name: string;
  cityId: string;
  latitude?: number;
  longitude?: number;
}